requirements.txt
acme==1.21.0
arrow==1.3.0
asgiref==3.7.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
binaryornot==0.4.4
blinker==1.4
boto3==1.28.62
botocore==1.31.62
Brotli==1.1.0
build==0.10.0
CacheControl==0.13.1
cachetools==5.5.2
certbot==1.21.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
cleo==2.0.1
click==8.0.3
colorama==0.4.4
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
cookiecutter==2.5.0
coverage==7.4.0
crashtest==0.4.1
cryptography==42.0.3
cssselect2==0.7.0
dateparser==1.1.8
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
Django==5.0.1
django-admin-extra-buttons==1.5.6
django-admin-rangefilter==0.12.0
django-admin-sortable2==2.1.10
django-adminactions==2.3.0
django-adminfilters==2.5.0
django-cleanup==8.0.0
django-constance==4.3.2
django-debug-toolbar==4.2.0
django-durationwidget==1.0.5
django-environ==0.11.2
django-filter==23.2
django-jsonform==2.23.1
django-nested-admin==4.0.2
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-tailwind==3.8.0
django-unfold==0.22.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dulwich==0.21.6
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==22.2.0
fastjsonschema==2.19.0
filelock==3.13.1
flake8==6.1.0
fontawesomefree==6.5.1
fonttools==4.56.0
freezegun==1.5.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.0.0
gspread==6.1.4
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.5
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
parsedatetime==2.6
pexpect==4.8.0
phonenumbers==8.13.27
phonetics==1.0.5
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.11.0
pycparser==2.21
pydyf==0.11.0
pyflakes==3.1.0
Pygments==2.17.2
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.3.0
pyOpenSSL==24.0.0
pyparsing==2.4.7
pyphen==0.17.2
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
python-gnupg==0.5.1
python-magic==0.4.24
python-monkey-business==1.0.0
python-slugify==8.0.1
pytz==2022.1
PyYAML==6.0.1
rapidfuzz==2.15.1
redis==5.2.1
regex==2023.10.3
requests==2.25.1
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
rich==13.7.0
rsa==4.9
s3transfer==0.7.0
SecretStorage==3.3.1
semver==2.13.0
sentry-sdk==1.40.4
service-identity==18.1.0
shellingham==1.5.3
six==1.16.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
ssh-import-id==5.11
tabulate==0.8.10
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.12.1
transitions==0.8.11
trove-classifiers==2023.8.7
Twisted==22.1.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.18
virtualenv==20.24.5
wadllib==1.3.6
weasyprint==64.1
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.1.7
xlwt==1.3.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
zopfli==0.2.3.post1
